# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soluct Test",
        "description": "Soluct Test API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List Products",
                "description": "Retrieves a list of all products.",
                "operationId": "98fdc6bfafda140e6760c7f569e359bc",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Status message for the product retrieval request.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Product"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product",
                "description": "Retrieves details of a single product.",
                "operationId": "e6fc1c399f951a5263dde00afb5e821d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Status message for the product retrieval request.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "product": {
                                                    "$ref": "#/components/schemas/Product"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Updates an existing product.",
                "operationId": "db3ff30f7d4dd14e19632c090d1d43bd",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "published"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://soluct.com.br"
                                    },
                                    "creator": {
                                        "type": "string",
                                        "example": "User XYZ"
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Sushi"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "brands": {
                                        "type": "string",
                                        "example": "Brand Name"
                                    },
                                    "categories": {
                                        "type": "string",
                                        "example": "Health"
                                    },
                                    "labels": {
                                        "type": "string",
                                        "example": "Food"
                                    },
                                    "cities": {
                                        "type": "string",
                                        "example": "Passo Fundo"
                                    },
                                    "purchase_places": {
                                        "type": "string",
                                        "example": "Shopping"
                                    },
                                    "stores": {
                                        "type": "string",
                                        "example": "Sushi House"
                                    },
                                    "ingredients_text": {
                                        "type": "string",
                                        "example": "Fish, Rice"
                                    },
                                    "traces": {
                                        "type": "string",
                                        "example": "Traces Text"
                                    },
                                    "serving_size": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "serving_quantity": {
                                        "type": "string",
                                        "example": "100g"
                                    },
                                    "nutriscore_score": {
                                        "type": "integer",
                                        "example": "32"
                                    },
                                    "nutriscore_grade": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "main_category": {
                                        "type": "string",
                                        "example": "Food"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://soluct.com/image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful product update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Status message for the product update request.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "product": {
                                                    "$ref": "#/components/schemas/Product"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "description": "Deletes a specified product.",
                "operationId": "ac55ba4e7659097f78152e5a28cffef2",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion of the product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Status message indicating the product was successfully deleted.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/products/sync": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Sync Products",
                "description": "Triggers a routine to import and sync products.",
                "operationId": "32cd47786712d8e7805abaebc02250eb",
                "responses": {
                    "200": {
                        "description": "Successful initiation of sync process",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Status message indicating the sync request was successful.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sync process failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "code": {
                        "description": "Code",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status",
                        "type": "string"
                    },
                    "imported_t": {
                        "description": "Imported date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "url": {
                        "description": "Url",
                        "type": "string"
                    },
                    "creator": {
                        "description": "Product Creator",
                        "type": "string"
                    },
                    "created_t": {
                        "description": "Created at",
                        "type": "integer"
                    },
                    "last_modified_t": {
                        "description": "Last modified at",
                        "type": "integer"
                    },
                    "product_name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "quantity": {
                        "description": "Quantity",
                        "type": "string"
                    },
                    "brands": {
                        "description": "Brands",
                        "type": "string"
                    },
                    "categories": {
                        "description": "Categories",
                        "type": "string"
                    },
                    "labels": {
                        "description": "Labels",
                        "type": "string"
                    },
                    "cities": {
                        "description": "Cities",
                        "type": "string"
                    },
                    "purchase_places": {
                        "description": "Purchase Places",
                        "type": "string"
                    },
                    "stores": {
                        "description": "Stores",
                        "type": "string"
                    },
                    "ingredients_text": {
                        "description": "Ingredients",
                        "type": "string"
                    },
                    "traces": {
                        "description": "Traces",
                        "type": "string"
                    },
                    "serving_size": {
                        "description": "Serving Size",
                        "type": "string"
                    },
                    "serving_quantity": {
                        "description": "Serving Quantity",
                        "type": "string"
                    },
                    "nutriscore_score": {
                        "description": "Nutriscore score",
                        "type": "integer"
                    },
                    "nutriscore_grade": {
                        "description": "Nutriscore grade",
                        "type": "string"
                    },
                    "main_category": {
                        "description": "Main category",
                        "type": "string"
                    },
                    "image_url": {
                        "description": "Image url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the auth token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}